Reorganize `event_api` module
