+++ optimizely/src/event_api/dispatcher.rs
@@ -1,0 +1,7 @@
+pub use batched_event_dispatcher::BatchedEventDispatcher;
+pub use event_dispatcher::EventDispatcher;
+pub use simple_event_dispatcher::SimpleEventDispatcher;
+
+mod batched_event_dispatcher;
+mod event_dispatcher;
+mod simple_event_dispatcher;
